refactor(sdk): move IndexedDB cache accessors to the @fluux/sdk/cache subpath (Phase 2)
The low-level message/avatar cache CRUD (~27 functions) sat on the curated
main entry, yet the stores are the source of truth and the write/delete ops
bypass their invariants — the exact 'apps that call saveMessage directly
bypass store invariants' hazard the review flagged. Only 8 of the 27 were
app-used, all reads.

- New `@fluux/sdk/cache` subpath (tsup entry + package export) documented as
  an advanced escape hatch; a guard test pins the accessors off the main
  entry and onto /cache.
- App's four legitimate consumers (reaction-notification message resolution,
  search context fetch, memory probe, clearLocalData) import from /cache; the
  reaction test mocks /cache instead of the main entry.

Main-entry surface shrinks by 27 exports (28 after the demo move).
