+++ apps/fluux/src/hooks/useReactionNotifications.ts
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next'
 import { useXMPP } from '@fluux/sdk'
 import { chatStore, roomStore, connectionStore, getBareJid, getLocalPart } from '@fluux/sdk'
 import { findMessageById } from '@fluux/sdk'
-import { getMessage as getCachedMessage, getMessageByStanzaId as getCachedMessageByStanzaId } from '@fluux/sdk'
-import { getRoomMessage as getCachedRoomMessage, getRoomMessageByStanzaId as getCachedRoomMessageByStanzaId } from '@fluux/sdk'
+import { getMessage as getCachedMessage, getMessageByStanzaId as getCachedMessageByStanzaId } from '@fluux/sdk/cache'
+import { getRoomMessage as getCachedRoomMessage, getRoomMessageByStanzaId as getCachedRoomMessageByStanzaId } from '@fluux/sdk/cache'
 import { useToastStore } from '@/stores/toastStore'
 import { useReactionMentionStore } from '@/stores/reactionMentionStore'
 import { useRouteSync } from '@/hooks'
